Fix: 빌드 오류 수정 완료 - node.js 모듈의 fs, path 등의 함수는 서버 사이드에서 실행되어야 한다 - 그러므로 이를 포함한 posts.ts 유틸모듈은 전부 서버 사이드에서만 실행되어야 함 - app 디렉토리 이외의 디렉토리에서 이 함수들이 호출되면 클라이언트 사이드에서 실행되는 것으로 간주해 빌드 오류가 발생한다 - 모든 데이터를 불러오는 로직은 app 디렉토리 하에서 실행해야 오류가 발생하지 않는다
