+++ package.json
@@ -29,7 +29,7 @@
         "glob": "^11.0.0",
         "gray-matter": "^4.0.3",
         "lucide-react": "^0.429.0",
-        "next": "14.2",
+        "next": "^14.2.6",
         "next-mdx-remote": "^5.0.0",
         "next-themes": "^0.3.0",
         "prettier": "^3.3.3",
